+++ efb_wechat_slave/wizard.py
@@ -144,6 +144,11 @@ flags_settings = {
          _('Choose the User Agent string to use when accessing Web Wechat. '
            'Leave undefined to use the default value provided by itchat.'
            )),
+    "text_post_processing":
+        (True, 'bool', None,
+         _('Determine whether to post-process text of messages received from '
+           'WeChat.'
+           )),
 }